Factor harness helpers into shared module
