+++ falcon_pachinko/testing/__init__.py
@@ -4,10 +4,14 @@ from __future__ import annotations
 
 import typing as typ
 
+from falcon_pachinko import _testing_harness as _testing_harness
+
 from ._common import MissingDependencyError
 from .client import TraceEvent, WebSocketTestClient
 from .harness import SimulatorConnection, SimulatorRouterHarness
 from .simulator import WebSocketSimulator
+
+websocket_simulator: typ.Any = getattr(_testing_harness, "websocket_simulator", None)
 
 __all__ = [
     "MissingDependencyError",
@@ -17,18 +21,3 @@ __all__ = [
     "WebSocketSimulator",
     "WebSocketTestClient",
 ]
-
-try:  # pragma: no cover - optional dependency for fixture registration
-    import pytest
-except ImportError:  # pragma: no cover - fixture only available under pytest
-    pytest = None  # type: ignore[assignment]
-else:
-
-    @pytest.fixture
-    def websocket_simulator() -> typ.Iterator[SimulatorRouterHarness]:
-        """Provide a router harness pre-wired with a simulator factory."""
-        harness = SimulatorRouterHarness()
-        try:
-            yield harness
-        finally:
-            harness._pending_simulator = None
